add richer location information to the position arg in compile-string-for-emacs
R's source location recording infrastructure works with line/column offsets
rather than with character offsets, so to get the right srcrefs associated
with some string source we need to communicate line and column information
from the source buffer.  Unfortunately, only the emacs side can do that -- in
principle, the R side has no direct access to the source file at all.  This
means a protocol change, converting the position argument in
swank:compile-string-for-emacs to a list of <position>s in the slime
source location grammar.  At present, pass in both :position and :line
kinds, and have swank:compile-string-for-emacs extract just the :position for
passing on to swank-compile-string.